python_version/requirements.txt
flask==3.0.2
pycryptodome==3.20.0
python-dotenv==1.0.1
pytest==8.0.2
flask-limiter==3.5.1 